refactor: add notifications as a module
