+++ src/modules/notifications/index.ts
@@ -1,6 +1,9 @@
 export {
   NotificationContextProvider,
   useNotificationsContext,
+  type NotificationPermissionStatus,
 } from './use-push-notifications';
 export {isConfigEnabled} from './utils';
 export {useOnPushNotificationOpened} from './use-on-push-notification-opened';
+export {type NotificationConfigGroup} from './types';
+export {useHasFareContractWithActivatedNotification} from './use-has-fare-contract-with-activated-notification';
