+++ src/onboarding/types.ts
@@ -1,7 +1,7 @@
 import {RootStackParamList} from '@atb/stacks-hierarchy';
 import {PermissionStatus} from 'react-native-permissions';
 import {AuthenticationType} from '@atb/auth/types';
-import {NotificationPermissionStatus} from '@atb/notifications/use-push-notifications';
+import {NotificationPermissionStatus} from '@atb/modules/notifications';
 import {MobileTokenStatus} from '@atb/mobile-token/types';
 
 type IsOnboardedStoreKey =
